RUN-18 Separate drag logic by HOC
