+++ bases/rsptx/assignment_server_api/assignment_builder/src/components/ui/EditableTable/hoc/withDragLogic.tsx
@@ -1,0 +1,39 @@
+import { ComponentType, useState } from "react";
+
+import { Nullable } from "@/types/common";
+import { DraggingExerciseColumns } from "@/types/components/editableTableCell";
+
+export interface WithDragLogicProps {
+  startRowIndex: Nullable<number>;
+  draggingFieldName: Nullable<DraggingExerciseColumns>;
+  handleMouseDown: (rowIndex: number, colKey: DraggingExerciseColumns) => void;
+  handleMouseUp: () => void;
+}
+
+export function withDragLogic<P extends WithDragLogicProps>(Component: ComponentType<P>) {
+  return function WrappedComponent(props: Omit<P, keyof WithDragLogicProps>) {
+    const [startRowIndex, setStartRowIndex] = useState<Nullable<number>>(null);
+    const [draggingFieldName, setDraggingFieldName] =
+      useState<Nullable<DraggingExerciseColumns>>(null);
+
+    const handleMouseDown = (rowIndex: number, colKey: DraggingExerciseColumns) => {
+      setStartRowIndex(rowIndex);
+      setDraggingFieldName(colKey);
+    };
+
+    const handleMouseUp = () => {
+      setStartRowIndex(null);
+      setDraggingFieldName(null);
+    };
+
+    return (
+      <Component
+        {...(props as P)}
+        startRowIndex={startRowIndex}
+        draggingFieldName={draggingFieldName}
+        handleMouseDown={handleMouseDown}
+        handleMouseUp={handleMouseUp}
+      />
+    );
+  };
+}
